scripts/advanced_requirements.txt
# Additional requirements for advanced dashboard
plotly>=5.15.0
networkx>=3.1
scipy>=1.10.0
scikit-learn>=1.3.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Core requirements (if not already installed)
streamlit>=1.25.0
pandas>=2.0.0
numpy>=1.21.0
biopython>=1.81
